完善连接图表代码。

+++ purenessscopeserver/purenessscopeserver/Common/define.h
@@ -1300,7 +1300,7 @@ private:
     unsigned long GetSystemTickCount()
     {
 #ifdef WIN32
-        return GetTickCount();
+        return (unsigned long)GetTickCount();
 #else
         struct timespec ts;
